Mithril Implimentation of React-Fractals
This converts from react to non-idomatic mithirl.  I'm doing some funky
stuff here to keep from storing all state on the vnode (keeping the
methods as standalone bound functions rather than living on
vnode.state).

This is probably not how I'd do it in a production app but I'm just
playing with it. :)